AD-HOC feat (service): Implement process separation
Currently, the thumbor process runs as a root level process on the
system. However, this poses additional risks in two components:

1. It is parsing what is arguably third party data. Though paths
   are controlled by proxying all connections via nginx, limiting
   the path there and listening only on local host it may still
   be possible for an attacker to access thumbor in an unexpected
   way with user supplied data, and
2. Thumbor is an unverified python system

As it stands any compromise of this process will lead to root level
control of the system.

Linux has the concept of "system users", which are users that are
designed explicitly to sandbox the application. This commit introduces a
user for thumbor, and modifies system init scripts to run thumbor as
this user.