AD-HOC feat (Thumbor): Add a configuration file
Thumbor allows configuring of a large number of parameters, outside
those supplied in the command line (and in /etc/defaults/thumbor). This
commits adds a configuration file generated by `thumbor-config`, but
does not configure it.

It further modifies the daemon to look for this configuration file.

The file itself has not been modified from that generated by
`thumbor-config`. Future work should build on this configuration to
allow modifications through Ansible variables.