AD-HOC feat (Thumbor): Implement the potential for optimizers
While thumbor can do arbitrary image translations and appears to be
doing some lossy image compression, it does not do optimisation as
efficiently as some tools expressly built for the purpose.

thumbor includes the idea of "optimziers"; programs run on the output of
thumbor that will further reduce file size. It has additionally
implemented jpegtran as one of these optimizers.

This commit allows the implementation of these optimizers, but does not
implement any specific optimizer.